+++ source/dotnet/domain-tests/DomainTests.csproj
@@ -10,6 +10,14 @@
     <Exec WorkingDirectory="$(ProjectDir)" EnvironmentVariables="ASPNETCORE_ENVIRONMENT=Development" Command="if not exist Clients/Wholesale/V3/swagger.json $(NSwagExe_Net60) run Clients/Wholesale/V3/nswag.json /variables:Configuration=$(Configuration)" Condition="!Exists('Clients/Wholesale/V3/swagger.json')" />
   </Target>
 
+  <ItemGroup>
+    <None Remove="Features\Calculations\TestData\amount_for_es_for_hourly_tarif_40000_for_e17_e02.csv" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <EmbeddedResource Include="Features\Calculations\TestData\amount_for_es_for_hourly_tarif_40000_for_e17_e02.csv" />
+  </ItemGroup>
+
 
   <ItemGroup>
     <PackageReference Include="Azure.Messaging.ServiceBus" Version="7.16.2" />
