+++ source/dotnet/domain-tests/Features/Calculations/TestData/amount_for_es_for_hourly_tarif_40000_for_e17_e02.csv
@@ -1,6 +1,6 @@
 grid_area;energy_supplier_id;quantity;time;price;amount;charge_code;
 804;5790001687137;54.658;2023-01-31 23:00:00;0.756998;41.375997;40000;
-804;5,79E+12;54.658;2023-02-01 00:00:00;0.756998;41.375997;40000;
+804;5790001687137;54.658;2023-02-01 00:00:00;0.756998;41.375997;40000;
 804;5790001687137;54.658;2023-02-01 01:00:00;0.756998;41.375997;40000;
 804;5790001687137;54.658;2023-02-01 02:00:00;0.756998;41.375997;40000;
 804;5790001687137;54.658;2023-02-01 03:00:00;0.756998;41.375997;40000;
